refactor: split cli commands into modules
